% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enigma_mdl.R
\name{enigma_mdl}
\alias{enigma_mdl}
\title{Isotope-genetics BMM}
\usage{
enigma_mdl(
  dat_in,
  nreps,
  nburn,
  thin,
  nchains,
  nadapt = 0,
  keep_burn = FALSE,
  cond_gsi = TRUE,
  file = NULL,
  seed = NULL,
  family = "normal"
)
}
\arguments{
\item{dat_in}{Input data (as a list object).}

\item{nreps}{Total number of iterations (includes burn-ins).}

\item{nburn}{Number of warm-up runs.}

\item{thin}{Frequency to thin the output.}

\item{nchains}{Number of independent MCMC processes.}

\item{nadapt}{Number of adaptation run (default is 0). Only available when
running model in fully Bayesian mode.}

\item{keep_burn}{Logical (default = FALSE). To save the burn-ins or not.}

\item{cond_gsi}{Logical (default = TRUE). To run the model in conditional GSI mode.}

\item{file}{File path to save the output in RDS file. Need to type out the full path and extension \code{.Rds}. Default = NULL for not saving the output.}

\item{seed}{Random seed for reproducibility. Default = NULL (no random seed).}

\item{family}{An option to model the isoscape in normal (Gaussian) or multinomial distribution. Default is "normal".}
}
\value{
A list containing:
\itemize{
\item Summary of the estimates
\item Trace of posterior samples
\item Individual assignment history
}
}
\description{
Run Bayesian mixture model (i.e., Pella-Madusa model) enhanced with isotope information.
}
\examples{
# prep input data
enigma_data <- prep_enigma_data(mixture_data = mix_iso, baseline_data = baseline, pop_info = ayk_pops60)

# run model
enigma_out <- enigma_mdl(enigma_data, 20, 10, 1, 3)

}
