% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enigma_data.R
\name{prep_enigma_data}
\alias{prep_enigma_data}
\title{Prepping isotope-genetics BMM input data}
\usage{
prep_enigma_data(
  mixture_data,
  baseline_data,
  pop_info,
  isoscape,
  file = NULL,
  loci = NULL
)
}
\arguments{
\item{mixture_data}{Mixture data in GCL or \pkg{rubias} format.}

\item{baseline_data}{Baseline data in GCL or \pkg{rubias} format.}

\item{pop_info}{Population information for the baseline. A tibble with columns
collection (collection names), repunit (reporting unit names),
grpvec (group numbers), origin (wild/hatchery).}

\item{isoscape}{Tibble of isoscape that is consisted of two columns:
\itemize{
\item \code{sr_mean} - mean Sr readings for each population of the genetic baseline,
\item \code{sr_sd} - standard deviation for Sr readings for each population of the genetic baseline.
}}

\item{file}{Where you want to save a copy of input data as a RDS file.
Need to type out full path and extension \code{.Rds}.
Leave it empty if you don't want to save a copy.}

\item{loci}{Optional. Provide loci for the mixture or baseline as a fail-safe check.}
}
\value{
A list objects as the input data for enigma_mdl()
}
\description{
Prep input data for Bayesian mixture model (i.e., Pella-Madusa model) enhanced with isotope information.
}
\examples{
# prep input data
enigma_data <- prep_enigma_data(mixture_data = mix_iso, baseline_data = baseline, pop_info = ayk_pops60)

}
