% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enigma_data.R
\name{prep_ichthy_data}
\alias{prep_ichthy_data}
\title{Prepping ichthy-genetics BMM input data}
\usage{
prep_ichthy_data(
  mixture_data,
  baseline_data,
  pop_info,
  ichthy_status,
  file = NULL,
  loci = NULL
)
}
\arguments{
\item{mixture_data}{Mixture data in GCL or \pkg{rubias} format.}

\item{baseline_data}{Baseline data in GCL or \pkg{rubias} format.}

\item{pop_info}{Population information for the baseline. A tibble with columns
\code{collection} (collection names), \code{repunit} (reporting unit names),
\code{grpvec} (group numbers), \code{origin} (wild/hatchery; optional).}

\item{ichthy_status}{Ichthyphonus information for mixture individuals. It is a tibble with two columns:
\itemize{
\item \code{indiv} - individual identification, the same as mixture data,
\item \code{ich} - ichthyphonus status, 1 = positive, 0 = negative.
}}

\item{file}{Where you want to save a copy of input data as a RDS file.
Need to type out full path and extension \code{.Rds}.
Leave it empty if you don't want to save a copy.}

\item{loci}{Optional. Provide loci for the mixture or baseline as a fail-safe check.}
}
\value{
A list objects as the input data for enigma_mdl()
}
\description{
Prep input data for Bayesian mixture model (i.e., Pella-Madusa model) integrated with ithchyphonus information.
}
\examples{
\dontrun{
# prep input data
ichthy_data <- prep_ichthy_data(mixture_data = mix, baseline_data = baseline, pop_info = ayk_pops60, ichthy_status = ichthy)
}

}
