% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enigma_data.R
\name{prep_enigma_data}
\alias{prep_enigma_data}
\title{Prepping isotope-genetics BMM input data}
\usage{
prep_enigma_data(
  mixture_data,
  baseline_data,
  pop_info,
  isobreaks = 5,
  isoscape = NULL,
  file = NULL,
  loci = NULL
)
}
\arguments{
\item{mixture_data}{Mixture data in GCL or \pkg{rubias} format with an additional column \code{sr_val} to specify isotope values if available.}

\item{baseline_data}{Baseline data in GCL or \pkg{rubias} format. It's optional to include a column \code{sr_val} for isotope values. Isotope value can be specified as a separate object \code{isoscape} (see below).}

\item{pop_info}{Population information for the baseline. A tibble with columns
collection (collection names), repunit (reporting unit names),
grpvec (group numbers), origin (wild/hatchery).}

\item{isobreaks}{Indicate where to cut and put isoscape values into bins, or how many bins the isoscape values should be grouped. Recommend manually set where to cut.}

\item{isoscape}{Option to provide isoscape if there are no isotope values (\code{sr_val}) for the baseline. Isoscape is a tibble with three columns:
\itemize{
\item \code{collection} - collection names,
\item \code{sr_mean} - mean Sr readings for each population of the genetic baseline,
\item \code{sr_sd} - standard deviation for Sr readings for each population of the genetic baseline.
}}

\item{file}{Where you want to save a copy of input data as a RDS file.
Need to type out full path and extension \code{.Rds}.
Leave it empty if you don't want to save a copy.}

\item{loci}{Optional. Provide loci for the mixture or baseline as a fail-safe check.}
}
\value{
A list objects as the input data for enigma_mdl()
}
\description{
Prep input data for Bayesian mixture model (i.e., Pella-Madusa model) enhanced with isotope information.
}
\examples{
\dontrun{
# prep input data
enigma_data <- prep_enigma_data(mixture_data = mix_iso, baseline_data = baseline, pop_info = ayk_pops60)
}

}
